Refactoring - component names
