+++ src/containers/BooksList.js
@@ -2,15 +2,15 @@ import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import axios from 'axios';
-import Bk from '../components/Bk';
+import Book from '../components/Book';
 import { getBooksByFilter } from '../selectors';
 import { registerBooks } from '../actions/index';
 import { httpProtocol, host, port } from '../envVariables';
 
 class BooksList extends React.Component {
   componentDidMount() {
-    const { props: { registerBooks, authentication } } = this;
-    axios.get(`${httpProtocol}://${host}:${port}/books`, { headers: { Authorization: `Bearer ${authentication.authentication_token}` } })
+    const { props: { registerBooks, user } } = this;
+    axios.get(`${httpProtocol}://${host}:${port}/books`, { headers: { Authorization: `Bearer ${user.authentication_token}` } })
       .then(response => {
         registerBooks(response.data.reverse());
       });
@@ -21,7 +21,7 @@ class BooksList extends React.Component {
     return (
       <div>
         {books && books.length ? (
-          books.map(book => <Bk key={`book-${book.id}`} book={book} showReading={showReading} />)
+          books.map(book => <Book key={`book-${book.id}`} book={book} showReading={showReading} />)
         ) : null}
       </div>
     );
@@ -29,16 +29,16 @@ class BooksList extends React.Component {
 }
 
 const mapStateToProps = state => {
-  const { filter, authentication } = state;
+  const { filter, user } = state;
   const books = getBooksByFilter(state, filter);
-  return { books, authentication };
+  return { books, user };
 };
 
 BooksList.propTypes = {
   books: PropTypes.arrayOf(PropTypes.object).isRequired,
   registerBooks: PropTypes.func.isRequired,
   showReading: PropTypes.func.isRequired,
-  authentication: PropTypes.objectOf(PropTypes.any).isRequired,
+  user: PropTypes.objectOf(PropTypes.any).isRequired,
 };
 
 export default connect(mapStateToProps, { registerBooks })(BooksList);
